% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-own.R
\name{collect_data_in_clone}
\alias{collect_data_in_clone}
\title{Automated data collection}
\usage{
collect_data_in_clone(refs = commandArgs(TRUE))
}
\arguments{
\item{refs}{\code{[character]}\cr Git references to collect benchmark data for}
}
\description{
Performs the following tasks:
\itemize{
\item Clones the \code{dplyr.benchmark} repository to a temporary directory
\item Collects benchmark data from the given references, by default taken from
command-line args
\item Commits and pushes
}
}

