% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{compute_calibrated_time}
\alias{compute_calibrated_time}
\title{Compute calibrated time}
\usage{
compute_calibrated_time(data)
}
\arguments{
\item{data}{A data frame with at least the columns \code{name} and \code{median_time}}
}
\description{
Median time divided by the time the \code{calibration} run takes.
}
