% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{detect_jumps}
\alias{detect_jumps}
\title{Detect jumps}
\usage{
detect_jumps(plot_data)
}
\arguments{
\item{plot_data}{A data frame as returned by \code{\link[=get_plot_data]{get_plot_data()}}}
}
\value{
\code{plot_data} augmented with columns provided by
\code{\link[tsoutliers:locate.outliers]{tsoutliers::locate.outliers()}}
}
\description{
Detect jumps
}

