% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{get_log_df}
\alias{get_log_df}
\title{Get a dataframe with dplyr log entries}
\usage{
get_log_df(ref = "master", repo = dplyr_repo())
}
\arguments{
\item{ref}{Git revision, default \code{master}}

\item{repo}{A git2r repository, default from GitHub}
}
\description{
Get a dataframe with dplyr log entries
}
