% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-own.R
\name{collect_data_in_clone}
\alias{collect_data_in_clone}
\title{Automated data collection}
\usage{
collect_data_in_clone(ref = "master", only_new = TRUE)
}
\arguments{
\item{ref}{\code{[character(1)]}\cr A Git refspec for dplyr revisions to test.
See help for \code{git parse-rev}, use \code{ref^!} to collect for a single revision.}

\item{only_new}{\code{[logical(1)]}\cr If \code{TRUE}, collect data for all references
up to the first for which data has been collected already.}
}
\description{
Performs the following tasks:
\itemize{
\item Clones the \code{dplyr.benchmark} repository to a temporary directory
\item Collects benchmark data from the given references, by default all
references in master that haven't been collected yet
\item Commits and pushes
}
}
